requirements.txt
Flask
requests
gunicorn
weasyprint
python-barcode
